+++ apidaora/app.py
@@ -1,9 +1,10 @@
 from collections import defaultdict
-from typing import DefaultDict, List, Optional, Sequence, Union
+from typing import Any, DefaultDict, List, Optional, Sequence, Type, Union
 
 from .asgi.app import asgi_app
 from .asgi.base import ASGIApp
 from .asgi.router import Controller, Route, make_router
+from .class_controller import ClassController
 from .controllers.background_task import BackgroundTask
 from .method import MethodType
 from .middlewares import Middlewares
@@ -14,9 +15,14 @@ from .route.factory import make_route
 Controllers = Union[
     Controller,
     BackgroundTask,
-    Sequence[Union[Controller, BackgroundTask]],
-    Sequence[Controller],
-    Sequence[BackgroundTask],
+    ClassController,
+    Type[ClassController],
+    Sequence[
+        Union[
+            Controller, BackgroundTask, ClassController, Type[ClassController],
+        ]
+    ],
+    Any,
 ]
 
 
@@ -26,6 +32,7 @@ def appdaora(
     options: bool = False,
 ) -> ASGIApp:
     routes = []
+    func_controllers: List[Union[Controller, BackgroundTask]] = []
     path_methods_map: DefaultDict[str, List[MethodType]] = defaultdict(list)
 
     def update_path_methods_map(routes: Sequence[Route]) -> None:
@@ -41,6 +48,19 @@ def appdaora(
         controllers = [controllers]
 
     for controller in controllers:
+        if (
+            isinstance(controller, ClassController)
+            or (
+                isinstance(controller, type)
+                and issubclass(controller, ClassController)
+            )
+            or (hasattr(controller, 'path'))
+        ):
+            routes.extend(get_class_routes(controller))
+        else:
+            func_controllers.append(controller)  # type: ignore
+
+    for controller in func_controllers:
         if isinstance(controller, BackgroundTask):
             routes.extend(controller.create.routes)
             routes.extend(controller.get_results.routes)
@@ -72,3 +92,18 @@ def appdaora(
             )
 
     return asgi_app(make_router(routes, middlewares=middlewares))
+
+
+def get_class_routes(controller: Any) -> List[Route]:
+    if isinstance(controller, type):
+        controller = controller()
+
+    return [
+        make_route(
+            path_pattern=controller.path,
+            method=method,
+            controller=getattr(controller, method_str),
+        )
+        for method in MethodType
+        if hasattr(controller, method_str := method.value.lower())  # noqa
+    ]