+++ docs/src/asgi_module/asgi_module.py
@@ -2,8 +2,7 @@ from typing import TypedDict
 
 from jsondaora import jsondaora, typed_dict_asjson
 
-from apidaora.core import appdaora_core, route
-from apidaora.core.responses import JSON_RESPONSE
+from apidaora.core import JSON_RESPONSE, appdaora_core, route
 
 
 @jsondaora
@@ -13,7 +12,7 @@ class You(TypedDict):
 
 
 @route.get('/hello/{name}', query=True)
-async def hello_controller(path_args, query_dict, headers, body):
+async def hello_controller(path_args, query_dict, headers, body):  # type: ignore
     you = You(name=path_args['name'], location=query_dict['location'][0])
     body = typed_dict_asjson(you, You)
     return JSON_RESPONSE, body