+++ docs/src/background_task_controller/background_task_controller.py
@@ -4,7 +4,7 @@ from apidaora import appdaora, route
 
 
 @route.background('/hello')
-def hello_task(name: str):
+def hello_task(name: str) -> str:
     time.sleep(1)
     return f'Hello {name}!'