+++ docs/src/redis_background_task_async_controller/redis_background_task_async_controller.py
@@ -8,7 +8,7 @@ REDIS_URI = f"redis://{os.environ.get('REDIS', '')}"
 
 
 @route.background('/hello', tasks_repository=REDIS_URI)
-async def hello_task(name: str):
+async def hello_task(name: str) -> str:
     await asyncio.sleep(1)
     return f'Hello {name}!'