+++ docs/src/redis_background_task_controller/redis_background_task_controller.py
@@ -8,7 +8,7 @@ REDIS_URI = f"redis://{os.environ.get('REDIS', '')}"
 
 
 @route.background('/hello', tasks_repository=REDIS_URI)
-def hello_task(name: str):
+def hello_task(name: str) -> str:
     time.sleep(1)
     return f'Hello {name}!'